stress-apparmor: size arg of read should be size_t, cast it so
Clean up warning on gcc + clang + pedantic build flags

+++ stress-apparmor.c
@@ -140,7 +140,7 @@ static void stress_apparmor_read(const char *path)
 redo:
 		if (!keep_stressing_flag() || !apparmor_run)
 			break;
-		ret = read(fd, buffer, sz);
+		ret = read(fd, buffer, (size_t)sz);
 		if (ret < 0) {
 			if ((errno == EAGAIN) || (errno == EINTR))
 				goto redo;
